moved app metrics repositories from /internal to /pkg
